requirements.txt
scipy
pandas
matplotlib
ipython
numpy
python-decouple